feat: add ability to style fieldset label like a heading-4
