+++ components/inputs/test/golden/input-fieldset/chromium/fieldset-manual.png
